requirements.txt
django==4.2.6
djangorestframework==3.14.0
psycopg2==2.9.9
drf-yasg==1.21.7
setuptools==68.2.2